Extract tracer tests that are Linux specific
Move sendfile test to linux test file
