+++ pkg/network/tracer/tracer_shared.go
@@ -1,0 +1,17 @@
+package tracer
+
+import (
+	"github.com/DataDog/datadog-agent/pkg/network"
+)
+
+// shouldSkipConnection returns whether or not the tracer should ignore a given connection:
+//  • Local DNS (*:53) requests if configured (default: true)
+func (t *Tracer) shouldSkipConnection(conn *network.ConnectionStats) bool {
+	isDNSConnection := conn.DPort == 53 || conn.SPort == 53
+	if !t.config.CollectLocalDNS && isDNSConnection && conn.Dest.IsLoopback() {
+		return true
+	} else if network.IsExcludedConnection(t.sourceExcludes, t.destExcludes, conn) {
+		return true
+	}
+	return false
+}
