+++ pkg/network/tracer/tracer_windows_test.go
@@ -1,0 +1,15 @@
+package tracer
+
+import "testing"
+
+func dnsSupported(t *testing.T) bool {
+	return true
+}
+
+func httpSupported(t *testing.T) bool {
+	return false
+}
+
+func connectionBufferCapacity(t *Tracer) int {
+	return t.connStatsActive.Capacity() + t.connStatsClosed.Capacity()
+}
